tests/wpt/tests/html/canvas/offscreen/transformations/2d.transformation.scale.negative.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.transformation.scale.negative</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.transformation.scale.negative</h1>
<p class="desc">scale() with negative scale factors works</p>


<script>
var t = async_test("scale() with negative scale factors works");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.fillStyle = '#f00';
  ctx.fillRect(0, 0, 100, 50);

  ctx.save();
  ctx.scale(-1, 1);
  ctx.fillStyle = '#0f0';
  ctx.fillRect(-50, 0, 50, 50);
  ctx.restore();

  ctx.save();
  ctx.scale(1, -1);
  ctx.fillStyle = '#0f0';
  ctx.fillRect(50, -50, 50, 50);
  ctx.restore();
  _assertPixel(canvas, 25,25, 0,255,0,255);
  _assertPixel(canvas, 75,25, 0,255,0,255);
  t.done();

});
</script>
